Refactor core.repository.ResourceForbiddenException to spi.AccessControlException
