+++ digdag-server/src/main/java/io/digdag/server/ServerModule.java
@@ -16,7 +16,6 @@ import io.digdag.core.database.ThreadLocalTransactionManager;
 import io.digdag.core.database.TransactionManager;
 import io.digdag.core.repository.ModelValidationException;
 import io.digdag.core.repository.ResourceConflictException;
-import io.digdag.core.repository.ResourceForbiddenException;
 import io.digdag.core.repository.ResourceLimitExceededException;
 import io.digdag.core.repository.ResourceNotFoundException;
 import io.digdag.guice.rs.GuiceRsModule;
@@ -32,6 +31,7 @@ import io.digdag.server.rs.SessionResource;
 import io.digdag.server.rs.UiResource;
 import io.digdag.server.rs.VersionResource;
 import io.digdag.server.rs.WorkflowResource;
+import io.digdag.spi.AccessControlException;
 import io.digdag.spi.AccessController;
 import io.digdag.spi.SecretControlStoreManager;
 import io.digdag.spi.SecretStoreManager;
@@ -131,7 +131,7 @@ public class ServerModule
     protected void bindExceptionhandlers(ApplicationBindingBuilder builder)
     {
         builder
-            .addProviderInstance(new GenericJsonExceptionHandler<ResourceForbiddenException>(Response.Status.FORBIDDEN) { })
+            .addProviderInstance(new GenericJsonExceptionHandler<AccessControlException>(Response.Status.FORBIDDEN) { })
             .addProviderInstance(new GenericJsonExceptionHandler<ResourceNotFoundException>(Response.Status.NOT_FOUND) { })
             .addProviderInstance(new GenericJsonExceptionHandler<StorageFileNotFoundException>(Response.Status.NOT_FOUND) { })
             .addProviderInstance(new GenericJsonExceptionHandler<ResourceConflictException>(Response.Status.CONFLICT) { })
